Create insertInternal Method

+++ src/main/java/edu/hanyang/submit/Block.java
@@ -1,7 +1,7 @@
 package edu.hanyang.submit;
 
 public class Block {
-    public long parent=0;
+    public int parent=0;
     public int type = 0; //1 is non-leaf, 0 is leaf
     public int nkeys = 0;
     private int max;
